# src/IO.Swagger/Models/PlaylistUpdate.cs
/*
 * Song, Artist, Album, and Playlist API
 *
 * API to manage songs, artists, albums, and playlists.
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class PlaylistUpdate : IEquatable<PlaylistUpdate>
    { 
        /// <summary>
        /// Gets or Sets Playlists
        /// </summary>

        [DataMember(Name="playlists")]
        public List<string> Playlists { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class PlaylistUpdate {\n");
            sb.Append("  Playlists: ").Append(Playlists).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((PlaylistUpdate)obj);
        }

        /// <summary>
        /// Returns true if PlaylistUpdate instances are equal
        /// </summary>
        /// <param name="other">Instance of PlaylistUpdate to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(PlaylistUpdate other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Playlists == other.Playlists ||
                    Playlists != null &&
                    Playlists.SequenceEqual(other.Playlists)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Playlists != null)
                    hashCode = hashCode * 59 + Playlists.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(PlaylistUpdate left, PlaylistUpdate right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(PlaylistUpdate left, PlaylistUpdate right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}